+++ feature/change-password/src/main/java/team/retum/jobis/change/password/navigation/ComparePasswordNavigation.kt
@@ -1,10 +1,10 @@
-package team.retum.jobis.change_password.navigation
+package team.retum.jobis.change.password.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import team.retum.common.enums.ResetPasswordNavigationArgumentType
-import team.retum.jobis.change_password.ui.ComparePassword
+import team.retum.jobis.change.password.ui.ComparePassword
 
 const val NAVIGATION_COMPARE_PASSWORD = "confirmPassword"
