+++ feature/change-password/src/main/java/team/retum/jobis/change/password/ui/ComparePasswordScreen.kt
@@ -1,4 +1,4 @@
-package team.retum.jobis.change_password.ui
+package team.retum.jobis.change.password.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -12,10 +12,10 @@ import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import team.retum.common.enums.ResetPasswordNavigationArgumentType
-import team.retum.jobis.change_password.R
-import team.retum.jobis.change_password.viewmodel.ComparePasswordSideEffect
-import team.retum.jobis.change_password.viewmodel.ComparePasswordState
-import team.retum.jobis.change_password.viewmodel.ComparePasswordViewModel
+import team.retum.jobis.change.R
+import team.retum.jobis.change.password.viewmodel.ComparePasswordSideEffect
+import team.retum.jobis.change.password.viewmodel.ComparePasswordState
+import team.retum.jobis.change.password.viewmodel.ComparePasswordViewModel
 import team.retum.jobisdesignsystemv2.appbar.JobisLargeTopAppBar
 import team.retum.jobisdesignsystemv2.button.ButtonColor
 import team.retum.jobisdesignsystemv2.button.JobisButton
